Move users to library
