+++ apps/api-gateway/src/app.module.ts
@@ -2,10 +2,9 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
-import { AuthModule, AuthGuard } from '@app/auth';
+import { AuthModule, AuthGuard, UserModule } from '@app/auth';
 import { APP_GUARD } from '@nestjs/core';
 import { KafkaModule } from '@app/kafka';
-import { UserModule } from './features/user/user.module';
 import { TargetModule } from './features/target/target.module';
 import { HttpModule } from '@nestjs/axios';
 import { SubmissionModule } from './features/submission/submission.module';
