+++ libs/auth/src/auth.controller.ts
@@ -2,7 +2,7 @@ import { Body, Controller, Post } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { UserDto } from './dto/user.dto';
 import { ApiBody, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
-import { Public } from '../../../apps/api-gateway/src/common/decorators/public';
+import { Public } from './decorators/public';
 
 @Controller('auth')
 @ApiTags('auth')
