+++ libs/auth/src/user/index.ts
@@ -1,0 +1,3 @@
+export * from './user.entity';
+export * from './user.service';
+export * from './user.module';
