+++ libs/auth/src/user/user.entity.ts
